requirements.txt
sanic~=20.12.0
Sanic-Cors~=0.10
sentry-sdk~=0.19.5
requests~=2.25.1
gql~=2.0.0
PyJWT~=2.0.0
Jinja2~=2.11